Replace a nice 17-line sh script with an equivalent 58-line python script.
To be fair, the python script has more documentation around it, and
will be easier to expand later.  And python style demands a lot of
empty lines.  But still, the bloat is definitely noticeable!  It's all
for a good cause.

Later, I will add in some of the nice functionality mentioned in the
docstring.