Revamp the name-matching to prefer full name-matches over substring matches.
Summary:
Now 'arc mail chris' will go to chris, and not ask if you really meant
"Cameron Christtensen."

Test Plan:
Ran
   arc diff --rr chris
and it did not prompt me to disambiguate.  Then I did
   arc diff --rr chri
and it said:
   Hmm...There are a few folks matching "chri":
     1. cameron (Cameron Christensen)
     2. chris (Chris Klaiber)
   Which "chri" did you mean?
Then I did
   arc diff --rr alpert
and it said:
   Hmm...There are a few folks matching "alpert":
     1. alpert (Ben Alpert)
     2. leah (Leah Alpert)
   Which "alpert" did you mean?
Then I did
   arc diff --rr alpert@
and it did not ask me to disambiguate.

Reviewers: chris, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D10514