Move everyone's .arcrc over to https.
Summary:
Now that nginx supports https communication to phabricator, this
should be the only step required before we turn on
security.require-https in the phabricator configs.

Test Plan:
Ran, in the arcanist repo-root:
   % python
   >>> execfile("khan-bin/arc")
   [ignored the IOError]
   >>> _update_arcrc(".")
Then examined my ~/.arcrc and saw it say:
    "default": "https://phabricator.khanacademy.org",
and have a new entry in "hosts".  Then I ran
    echo '{}' | arc call-conduit repository.query
and saw it give me some json output that had a bunch of repo info in it.

Reviewers: chris, alpert, benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D14708