Sync arc and arcrc even if we are installing a certificate
Summary:
If we are installing a certificate, we can't yet do most of our magic.
But we can still make sure we have latest arc (hopefully we do, since we
probably just cloned it) and initialize the user's `arcrc` (now is a
great time to do so in fact).  Let's do that, so setting up arc can be
just one command.

Test Plan:
- run `rm /tmp/arc.pull ~/.arcrc`, then
  `arc install-certificate https://phabricator.khanacademy.org`
  should work and install both my cert and the standard parts of
  `~/.arcrc`.

Reviewers: csilvers, jacqueline

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D45131