Allow batch linting of files in ScriptAndRegexLinter
Summary:
This commit addresses Jira ticket: INFRA-1462

The typical usage of the ScriptAndRegexLinter (and in fact all linters from Arcanist) requires us to lint one file at a time. However this is assuming you are configuring several linters directly through .arclint, which we are not. Since Khan uses its own custom linter (and therefore only ever needs one linter called on each file -- ka-lint), it is faster for us to call this on files in batch.

This change was actually more straightforward than originally expected, and only requires a new config to be added to the .arclint, which declares batch = true. With this flag, the script file of the script and regex is called in batch, currently on chunks of 32 files, as that is the predefined size of file chunking done by the built in ArcanistLintEngine when executing a linter. There is a TODO to consider allowing us to define other batch sizes if that makes sense, tho this would likely need to be a batch size that is a even divisor of 32. Other batch sizes or something bigger than 32 would require much work on arcanist than just amending the ScriptAndRegexLinter, and would be less than ideal.

For backwards compatibility with all existing repositories that use the ScriptAndRegexLinter, this new batch config is optional and will default to false if not provided in the .arclint file.

Test Plan:
To ensure linters are working as expected:
1. Add `batch: true` to webapp's .arclint file
2. Change handful of files in webapp
3. Run `arc lint`
4. Note lint failures displaying properly
5. Repeat for another repo which has .arclint configured (i.e. khan-linter)

To verify timing improvements:
1. Changed 5-10 files in webapp
2. Run 5 trials of `time arc lint` when batch: true in .arclint, then change to batch: false and do 5  more trials

Avg of trials (5 files):
Batched -> 10.383s
Non-batched -> 16.0474s
Improvement of ~35%.

Avg of trials (10 files):
Batched -> 15.936s
Non-batched -> 27.2056s
Improvement of ~41%.

3. Rerun the above with js changes

Avg of trials (5 files):
Batched -> 9.543s
Non-batched -> 13.147s
Improvement of ~27%.

Avg of trials (10 files):
Batched -> 12.5654s
Non-batched -> 19.3076s
Improvement of 35%.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D46540