Fix arc so `git rebase -x 'arc amend'` works even if we pull
Summary:
This fixes a bug that's been bugging me for years: if you call out to
arc as a `git rebase` exec line, and arc happens to decide to pull, it
crashes (with "not on a branch").  This appears to be because such
commands run with GIT_DIR and GIT_WORK_TREE set to the current repo,
which overrides the default of the current working directory, so they
end up operating on the repo you're rebasing, which sure isn't what you
want.  Now, we clear them about before pulling.

Test Plan: in another repo, `rm /tmp/arc.pull ; git rebase -x 'arc amend'`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D54785