Catch a new exception caused by not having an ssh key.
New users were getting this error:
```
[3:56 PM] Traceback (most recent call last):
  File "/Users/stanleyyip/khan/devtools/arcanist/khan-bin/arc", line 609, in <module>
    main(sys.argv[1:])
  File "/Users/stanleyyip/khan/devtools/arcanist/khan-bin/arc", line 582, in main
    _sync_arc(arc_root)
  File "/Users/stanleyyip/khan/devtools/arcanist/khan-bin/arc", line 167, in _sync_arc
    _get_user_info(arc_root, use_cache=False)   # updates the cache
  File "/Users/stanleyyip/khan/devtools/arcanist/khan-bin/arc", line 231, in _get_user_info
    retval = [(ui['userName'], ui['realName']) for ui in user_info
TypeError: 'NoneType' object is not iterable
```

Presumably this is due to the new arc changing its conduit output in a
way we're not noticing anymore.  I don't know how to repro so I'm just
catching the type error as well, with a TODO to convert it into a
RuntimeError one day.

Auditors: benkraft

Test plan:
Ran `arc which` without error.