Fix substring matching on projects.
Summary:
In theory if you say `arc mail --rr #foo` it's supposed to be a
substring match that matches a project like `#content-foo`.  But that
was broken because we weren't stripping off the `#`, so we were really
only doing a prefix match.

Issue: none

Test Plan:
I ran
   arc mail --rr #platform
on this diff and it properly showed `#infrastructure-platform` (as
well as `#ContentPlatform`) as opposed to saying no projects matched.

Reviewers: #infrastructure-platform, benkraft

Reviewed By: #infrastructure-platform, benkraft

Differential Revision: https://phabricator.khanacademy.org/D66192